Make the linter use a JAXB jar instead of the deprecated module.
The linter loads JAXB classes (in the javax.xml.bind package) at
runtime. Prior to this change, it was getting them by using
--add-modules=java.xml-bind. This works with an OpenJDK 9 toolchain,
but that module is deprecated, and it is removed in OpenJDK 11, so
this approach doesn't work with an upgraded toolchain. This change
drops the --add-modules option, and instead adds a JAR containing the
JAXB classes. This works on all toolchain versions.

Test: prebuilts/devtools/tools/lint packages/apps/Contacts
Bug: 135179316
Change-Id: I8ee090152e75a6d0f7fd89bdb95a70539cd556a0